requirements.txt
pyyaml~=6.0.3
numpy~=2.2.6